requirements.txt
moviepy==1.0.3
openai_whisper==20231117
pyperclip==1.8.2
